Title Screen Picture and Throwing Animation (#8)
* Title Screen Picture and Throwing Animation

I have added my completed illustration to the title screen, and I have added an animation that will play when the character is meant to be throwing a projectile. I haven't gotten the character to flip based on whether he's facing right or left when he's doing the throwing animation, though. I've also had an issue with the throwing animation looping after I press 'Q'

* Merge conflict resolved

---------

Co-authored-by: Elisha Lucile O <lucile.ong@my.utsa.edu>